re-structure code for PatientDetail
